penambahan surat

+++ app/Http/Controllers/DaftarMahasiswaController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\DaftarMahasiswa;
 use App\Http\Requests\StoreDaftarMahasiswaRequest;
 use App\Http\Requests\UpdateDaftarMahasiswaRequest;
+use App\Models\PengajuanSurat;
 use App\Models\User;
 use Illuminate\Http\Request;
 
@@ -36,14 +37,21 @@ class DaftarMahasiswaController extends Controller
             'name' => 'required',
             'username' => 'required|unique:users,username',
             'email' => 'required|email|unique:users,email',
+            'no_hp' => 'required',
             'password' => 'required|min:8',
-        ]);
+            'fakultas' => 'required',
+            'prodi' => 'required',
+        ],);
+
         User::create([
             'username' => $request->username,
             'name' => $request->name,
             'email' => $request->email,
+            'no_hp' => $request->no_hp,
             'role' => 'mahasiswa',
             'password' => bcrypt($request->password),
+            'fakultas' => $request->fakultas,
+            'prodi' => $request->prodi,
         ]);
         return redirect()->route('daftar-mahasiswa.index')->with('success', 'Mahasiswa berhasil ditambahkan');
     }
@@ -65,7 +73,10 @@ class DaftarMahasiswaController extends Controller
         $request->validate([
             'name' => 'required',
             'username' => 'required|unique:users,username,' . $id,
+            'no_hp' => 'required',
             'email' => 'required|email|unique:users,email,' . $id,
+            'fakultas' => 'required',
+            'prodi' => 'required',
         ]);
 
         $user = User::find($id);
@@ -73,6 +84,9 @@ class DaftarMahasiswaController extends Controller
         $updateData = [
             'username' => $request->username,
             'name' => $request->name,
+            'no_hp' => $request->no_hp,
+            'fakultas' => $request->fakultas,
+            'prodi' => $request->prodi,
             'email' => $request->email,
         ];
 
@@ -91,6 +105,15 @@ class DaftarMahasiswaController extends Controller
     public function destroy($id)
     {
         $user = User::find($id);
+        // Check if the user has any related PengajuanSurat records
+        $suratCount = PengajuanSurat::where('user_id', $id)->count();
+
+        if ($suratCount > 0) {
+            // If related records exist, redirect back with an error message
+            return redirect()->route('daftar-mahasiswa.index')->with('error', 'Mahasiswa tidak dapat dihapus karena memiliki pengajuan surat.');
+        }
+
+        // If no related records exist, proceed with deletion
         $user->delete();
         return redirect()->route('daftar-mahasiswa.index')->with('success', 'Mahasiswa berhasil dihapus');
     }
